SWC-6557: improvements per PR
